Renamed pipes to search-highlight
