--- src/app/pipes/search-highlight.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-import { SearchHighlight } from './search-highlight.pipe';
-
-@NgModule({
-  declarations: [ SearchHighlight ],
-  exports: [ SearchHighlight ]
-})
-export class SearchHighlightModule { }
