+++ src/app/search-highlight/examples/search-highlight-example.component.html
@@ -24,10 +24,10 @@
   <div>
     <tabset>
       <tab heading="html">
-        <include-content src="/src/app/pipes/examples/search-highlight-example.component.html"></include-content>
+        <include-content src="/src/app/search-highlight/examples/search-highlight-example.component.html"></include-content>
       </tab>
       <tab heading="typescript">
-        <include-content src="/src/app/pipes/examples/search-highlight-example.component.ts"></include-content>
+        <include-content src="/src/app/search-highlight/examples/search-highlight-example.component.ts"></include-content>
       </tab>
     </tabset>
   </div>
