+++ src/app/search-highlight/examples/search-highlight-example.component.ts
