+++ src/app/search-highlight/examples/search-highlight-example.module.ts
