+++ src/app/search-highlight/search-highlight.module.ts
@@ -1,0 +1,8 @@
+import { NgModule } from '@angular/core';
+import { SearchHighlightPipe } from './search-highlight.pipe';
+
+@NgModule({
+  declarations: [ SearchHighlightPipe ],
+  exports: [ SearchHighlightPipe ]
+})
+export class SearchHighlightModule { }
