+++ src/app/search-highlight/search-highlight.pipe.spec.ts
@@ -1,8 +1,8 @@
-import { SearchHighlight } from './search-highlight.pipe';
+import { SearchHighlightPipe } from './search-highlight.pipe';
 
 describe('Search highlight pipe', () => {
   // This pipe is a pure, stateless function so no need for BeforeEach
-  let pipe = new SearchHighlight();
+  let pipe = new SearchHighlightPipe();
 
   it('should transforms "Sudipta" to "S<b>ud</b>ipta"', () => {
     expect(pipe.transform('Sudipta', 'ud')).toBe('S<b>ud</b>ipta');
